renamed review-tutorial to getTutorialByUsername
